refactor: 에러폴백 일러스트 컴포넌트 공컴으로 이동
