+++ src/shared/components/errorFallback/ErrorImageGroup.css.ts
